@@ -6,26 +6,29 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  get_it: ^7.2.0
-  provider: ^6.0.2
-
-  intl: ^0.18.1
-  shared_preferences: ^2.0.13
-  flutter_svg: ^2.0.9
-
-  cupertino_icons: ^1.0.2
+  get_it: ^9.1.1
+  mobx: ^2.2.0+1
+  flutter_mobx: ^2.1.0+1
+  go_router: ^17.0.0
+  dio: ^5.7.0
+  shared_preferences: ^2.3.3
+  intl: ^0.20.2
+  flutter_svg: ^2.0.15
+  cupertino_icons: ^1.0.8
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^1.0.0
+  flutter_lints: ^6.0.0
+  build_runner: ^2.4.13
+  mobx_codegen: ^2.5.0+1
 
 flutter:
   uses-material-design: true
